sync-jobs: make Jenkins uses GitHub repos directly
Jenkins debian-glue plugin was not able to use the cloned repository on
the server since they were missing some refs. The Jenkins jobs are now
setup to use the GitHub repository directly. This has also the
advantage to let maintainer launch builds as soon as commits are pushed
to Github. They do not have to wait the sync utility to run (and fetch
latest commits) before launching the builds.