Add missing metadata for repos conf
This way debian-debian-glue (with its script generate-reprepro-codename)
is able to dynamically generate the APT repository distributions
configuration with all required metadata. With this feature, the admins
do no longer have to edit this file manually to add missing parameters.

NB: this requires a modified version of jenkins-debian-glue. The patch
has been submitted to upstream devs through the following PR:

    https://github.com/mika/jenkins-debian-glue/pull/189

By now, it is still pending for merge.