Introduce aspects jobs
Conventions for aspects packages Git repositories are defined on
scibian.org website:

  https://scibian.org/docs/aspects.html

This commit introduces the management of these jobs and their
specificities.

Unfortunately, Jenkins API (or jenkins-job-builder?) does not support
the + character in job names. Then, the notion of escaped codename has
been introduced to replace the + with _ in job names. This notably
affects jobs templates quite significantly.