sync-jobs: rename template after project dimension
Recent versions of jenkins-job-builder check if the dimension name of
multi-dimensional projects is included in job template name, or ignore
it otherwise.

In our case, we had a distros dimension but the template names did not
contain distros. Then, the distros dimension was not expanded by the
new version of jenkins-job-builder recently installed on scibian.org
server.

This is fixed by renaming the job templates to include the distros
dimension. Also, the template of project is updated to use the escaped
distribution name (used by aspects) as keys for the distros dimension.