sync-jobs: use http URI for local repo
The jenkins job template was using a file: URI for accessing the local
repository. It has worked sometime ago but pbuilder now complains it is
not able to find the file anymore. I'm not sure where the regression
comes from, maybe something more deeply chrooted since the server
upgrade to stretch? Or not binded in the chroot anymore? Anyway, using
the http URI seems OK.